requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.1.0.post1
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
alembic==1.13.1
bcrypt==4.1.2
python-dotenv==1.0.1
itsdangerous==2.2.0
Flask-Limiter==3.8.0
qrcode==7.4.2
reportlab==4.0.8
pikepdf==8.15.1
Pillow==10.3.0
pytesseract==0.3.10
minio==7.2.7
